third_party/blink/web_tests/platform/generic/http/tests/webfont/fallback-font-while-loading-expected.txt

This test checks that the fallback font is used for layout while a webfont is loading.

Target:
A text to be measured.
A text to be measured.
A text to be measured.
A text to be measured.
Reference:
A text to be measured.
A text to be measured.
A text to be measured.
A text to be measured.
arial
PASS: The width of target text and reference text is the same.
serif
PASS: The width of target text and reference text is the same.
default
PASS: The width of target text and reference text is the same.
